Split DeveloperMode into smaller components Add String and number conversion tools
